Fixing signal_loopbreak: eventmachine not initialized
Since we want to talk with EM the reactor should be running
first. This error occured randomly cause on every run
MiniTest has a different test order. So in the past it was
possible that EM was already initialized probably. I am not
sure whether that is the final fix or if I missed a test part

Signed-off-by: Lukas Matt <lukas@zauberstuhl.de>